assignment_1/images-classification/food101_project/requirements.txt
torch
torchvision
scikit-learn
pandas
matplotlib
Pillow
tqdm
gradio
